Add comprehensive lint configuration (Phase 4.11)
Implement recommended Rust and Clippy lints in Cargo.toml for improved
code quality and maintainability.

## Lints Added

### Rust Standard Lints
- ambiguous_negative_literals: warn
- redundant_imports: warn
- redundant_lifetimes: warn
- unsafe_op_in_unsafe_fn: warn
- unused_lifetimes: warn
- missing_debug_implementations: allow (not needed for UI framework)
- trivial_numeric_casts: allow (explicit casts aid clarity)

### Clippy Lint Groups
- cargo, complexity, correctness: warn
- pedantic, perf, style, suspicious: warn

### Clippy Restriction Lints (21 added)
- allow_attributes_without_reason
- as_underscore
- assertions_on_result_states
- clone_on_ref_ptr
- undocumented_unsafe_blocks
- and 16 more for enhanced code quality

### Pragmatic Allows
Configured allows for pedantic lints that are too strict for UI code:
- similar_names (x/y, w/h are clear in context)
- collapsible_if (separate ifs can be more readable)
- redundant_else (explicit else improves clarity)
- unreadable_literal (hex/binary literals don't need separators)
- module_name_repetitions (button::Button is idiomatic)
- must_use_candidate (too many false positives)

## Code Fixes
- Fixed redundant import of SHADOW_SIZE in view.rs
- Removed redundant i16 cast in window.rs (width/height already i16)

## Results
- All 180 tests passing ✅
- Only 5 expected warnings (intentionally unused code for future use)
- Build clean with comprehensive lint coverage

## Benefits
- Catches potential bugs early
- Enforces consistent code style
- Prevents regressions
- Guides contributors to best practices

+++ src/views/window.rs
@@ -209,7 +209,7 @@ impl View for Window {
                 // Update bounds (maintaining size)
                 let width = self.bounds.width();
                 let height = self.bounds.height();
-                self.bounds = Rect::new(new_x, new_y, new_x + width as i16, new_y + height as i16);
+                self.bounds = Rect::new(new_x, new_y, new_x + width, new_y + height);
 
                 // Update frame and interior bounds
                 self.frame.set_bounds(self.bounds);
